+++ demo/src/main/java/com/example/storage/product/ProductStorage.java
@@ -5,5 +5,5 @@ import com.example.entities.Product;
 public interface ProductStorage {
 	Product get(int ID) throws Exception;
 	Product[] getAll() throws Exception;
-	Product insert(String name, float price) throws Exception;
+	Product insert(String name, int price) throws Exception;
 }
